tensorflow/lite/micro/tools/make/downloads/cmsis/CMSIS/Core/Test/requirements.txt
# -*- coding: utf-8 -*-
#
# Python requirements for build.py script
#
python-matrix-runner~=1.2
lit~=17.0
